Pruthvi

+++ Account/views.py
@@ -20,7 +20,6 @@ from reportlab.lib.pagesizes import letter
 from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
 from reportlab.platypus import SimpleDocTemplate, Paragraph
 from Account.utils import decrypt_email, encrypt_email
-import requests
 import traceback
 from rest_framework.views import APIView
 from rest_framework import status
@@ -41,6 +40,8 @@ from django.db import IntegrityError
 from django.urls import reverse
 from django.http import HttpResponseBadRequest
 import logging
+import requests
+from django.db import models
 
 # Set up logging
 logger = logging.getLogger(__name__)
@@ -646,33 +647,33 @@ def tables(request):
 def citizenLoginAccount(request):
     try:
         if request.method == "GET":
-            request.session.flush()
             service_db = request.GET.get('service_db', 'default')
-            # service_db = request.GET.get('service_db', 'default') 
             request.session['service_db'] = service_db
             
             return render(request, 'citizenAccount/citizenLogin.html')
 
         elif request.method == "POST":
             service_db = request.session.get('service_db', 'default')
-            
             phone_number = request.POST.get('username', '').strip()
 
             if phone_number:
-                user = CustomUser.objects.get(phone=phone_number)
-                
-                if user:  
+                try:
+                    CustomUser.objects.get(phone=phone_number)
                     request.session['phone_number'] = phone_number
                     return redirect('OTPScreen')
-
-    except CustomUser.DoesNotExist:
-        messages.warning(request, "The phone number entered is not registered. Please register yourself.")
-        return redirect('citizenRegisterAccount')
+                except CustomUser.DoesNotExist:
+                    messages.warning(request, "The phone number entered is not registered. Please register yourself.")
+                    return redirect('citizenRegisterAccount')
 
     except Exception as e:
-        logger.error(f"An error occurred in citizenLoginAccount: {str(e)}", exc_info=True)
-        messages.error(request, "An unexpected error occurred. Please try again.")
+        tb = traceback.extract_tb(e.__traceback__)
+        fun = tb[0].name
+        callproc("stp_error_log",[fun,str(e),user])  
+        messages.error(request, 'Oops...! Something went wrong!')
+        # logger.error("An error occurred in citizenLoginAccount: %s", str(e), exc_info=True)
+        # messages.error(request, "An unexpected error occurred. Please try again.")
         return redirect('citizenRegisterAccount')
+        
 
     return render(request, 'citizenAccount/citizenLogin.html')
 
@@ -698,7 +699,7 @@ def citizenRegisterAccount(request):
         mobile_number = request.POST.get('mobileNumber').strip()
 
         if CustomUser.objects.filter(phone=mobile_number).exists():
-            messages.warning(request, "This mobile number is already registered. Please use a different number.")
+            messages.warning(request, "This mobile number is already registered. Please LogIn.")
             return redirect('citizenRegisterAccount')  
         else:
             request.session['first_name'] = first_name
@@ -716,40 +717,118 @@ def OTPScreen(request):
         request.session['service_db'] = service_db
 
         phone_number = request.session.get('phone_number')
-
+        sms_templates = smstext.objects.all()
+        
         try:
             otp = ''.join([str(random.randint(0, 9)) for _ in range(6)])
+            
             OTPVerification.objects.create(
                 mobile=phone_number,
                 otp_text=otp,
                 created_at=timezone.now()
             )
+            
+            if sms_templates:
+                template_id = sms_templates[0].template_id 
+                message = sms_templates[0].template_name 
+                
+                action = "Login"  
+                service = "Drainage connection"  
+
+                message = format_message(message, otp, action, service)
+                send_sms(phone_number, message, template_id)
+            
             messages.success(request, "OTP sent successfully!")
         except Exception as e:
             messages.error(request, "Failed to send OTP. Please try again.")
 
         return render(request, 'OTPScreen/OTPScreen.html')
 
+def format_message(template, otp_value, action, service):
+    message = template.replace("@otp", otp_value)
+    message = message.replace("@action", action)
+    message = message.replace("@service", service)
+    return message
+
+def send_sms(mobile, message, template_id):
+    try:
+        url = (
+            f"https://push3.aclgateway.com/servlet/com.aclwireless.pushconnectivity.listeners.TextListener"
+            f"?appid=MahaITcidc&userId=MahaITcidc&pass=mitcidc_10&contenttype=1"
+            f"&from=MAHGOV&to={mobile}&text={message}&alert=1&selfid=true&dlrreq=true"
+            f"&intflag=false&dtm={template_id}"
+        )
+        response = requests.get(url)
+        print(response.text)
+        
+        sms_log(response, mobile, message, template_id)
+        
+    except Exception as e:
+        messages.error("Failed to send OTP. Please try again.")
+
+# Optimized SMS log function
+def sms_log(response, mobile, message, template_id):
+    try:
+        content = response.text or ""
+        content_type = response.headers.get('Content-Type', '')
+        response_status = response.status_code or ""
+        is_successful = response.ok
+        response_url = getattr(response, 'url', '')
+        status_description = response.reason or str(response.status_code)
+        unique_id = '' 
+        user_id = '1' 
+
+        logging.info(f"SMS sent to {mobile} with status {response_status} and content: {content}")
+
+        smslog.objects.create(
+            mobile=mobile or '',
+            message=message or '',
+            user_id=user_id,
+            content=content,
+            status=response_status,
+            unique_id=unique_id,
+            content_type=content_type,
+            response_status=response_status,
+            is_successful=is_successful,
+            response_url=response_url,
+            status_description=status_description,
+            template_id=template_id or ''
+        )
+    except Exception as e:
+        tb = traceback.extract_tb(e.__traceback__)
+        fun = tb[0].name
+        callproc("stp_error_log",[fun,str(e),user])  
+        messages.error('Oops...! Something went wrong!')
+                
 @csrf_exempt
 def OTPScreenPost(request):
     if request.method == "POST":
         phone_number = request.session.get('phone_number')  
-        entered_otp = request.POST.get('OTPHere')
+        entered_otp = request.POST.get('otp')
 
         if not phone_number:
             messages.error(request, "Phone number not found. Please log in again.")
             return redirect('citizenLoginAccount')
 
         try:
-            otp_record = OTPVerification.objects.get(mobile=phone_number, otp_text=entered_otp)
-            otp_record.delete()
+            otp_record = OTPVerification.objects.filter(mobile=phone_number).order_by('-id').first()
+            
+            if otp_record and otp_record.otp_text == entered_otp:
+                otp_record.delete()
 
-            request.session['phone_number'] = phone_number
-            messages.success(request, "OTP verified successfully!")
-            return redirect('applicationFormIndex')
+                user = CustomUser.objects.get(phone=phone_number)
+                request.session['full_name'] = user.full_name
+                
+                messages.success(request, "OTP verified successfully!")
+                base_url = reverse('applicationFormIndex')
+                return redirect(f'{base_url}?service_db=1')
+                # return redirect('applicationFormIndex')
+            else:
+                messages.error(request, "Invalid OTP. Please try again.")
+                return redirect('citizenLoginAccount')
 
         except OTPVerification.DoesNotExist:
-            messages.error(request, "Invalid OTP. Please try again.")
+            messages.error(request, "No OTP record found. Please request a new OTP.")
             return redirect('citizenLoginAccount')
 
     return render(request, 'citizenAccount/citizenLogin.html', {'error': 'Invalid request method.'})
@@ -783,20 +862,39 @@ def OTPScreenRegistration(request):
         request.session['service_db'] = service_db
         
         phone_number = request.session.get('mobile_number')
-
+        sms_templates = smstext.objects.all()
+        
         if not phone_number:
             messages.error(request, "Phone number is required.")
             return redirect('citizenRegisterAccount')
 
         otp = ''.join([str(random.randint(0, 9)) for _ in range(6)])
-        
-        otp_record = OTPVerification.objects.create(
-            mobile=phone_number,
-            otp_text=otp,
-            created_at=timezone.now()
-        )
 
-        messages.success(request, "OTP sent successfully!")
+        try:
+            OTPVerification.objects.create(
+                mobile=phone_number,
+                otp_text=otp,
+                created_at=timezone.now()
+            )
+            
+            if sms_templates:
+                template_id = sms_templates[0].template_id
+                message = sms_templates[0].template_name
+                
+                action = "Registration"
+                service = "Drainage connection"
+                
+                message = format_message(message, otp, action, service)
+                
+                send_sms(phone_number, message, template_id)
+            
+            messages.success(request, "OTP sent successfully!")
+        except Exception as e:
+            tb = traceback.extract_tb(e.__traceback__)
+            fun = tb[0].name
+            callproc("stp_error_log", [fun, str(e), user])  
+            messages.error(request, f"Failed to send OTP. Error: {str(e)}")
+        
         return render(request, 'OTPScreen/OTPScreenRegistration.html')
 
 @csrf_exempt
@@ -806,32 +904,52 @@ def verify_otp(request):
         first_name = request.session.get('first_name')
         last_name = request.session.get('last_name')
         email = request.session.get('email')
-        entered_otp = request.POST.get('OTPHere')
+        entered_otp = request.POST.get('otp')
         
         if not phone_number:
             messages.error(request, "Phone number not found. Please log in again.")
             return redirect('citizenRegisterAccount')
 
         try:
-            otp_record = OTPVerification.objects.get(mobile=phone_number, otp_text=entered_otp)
 
-            otp_record.delete()
+            otp_record = OTPVerification.objects.filter(mobile=phone_number).order_by('-id').first()
             
-            user = CustomUser.objects.create(
-                full_name=f"{first_name} {last_name}",
-                email=email,
-                phone=phone_number,
-                first_time_login=1  
-            )
-            
-            request.session['user_id'] = user.id  # Store the newly created user's ID
-            request.session['phone_number'] = phone_number
+            if otp_record and otp_record.otp_text == entered_otp:
 
-            messages.success(request, "Registered successfully! OTP verified.")
-            return redirect('citizenLoginAccount')  
+                otp_record.delete()
+
+                user = CustomUser.objects.create(
+                    full_name=f"{first_name} {last_name}",
+                    email=email,
+                    phone=phone_number,
+                    first_time_login=1 
+                )
+
+                request.session['user_id'] = user.id
+                request.session['phone_number'] = phone_number
+
+                messages.success(request, "Registered successfully! OTP verified.")
+                return redirect('citizenLoginAccount')
+
+            else:
+                messages.error(request, "Invalid OTP. Please try again.")
+                context = {
+                    'firstName': first_name,
+                    'lastName': last_name,
+                    'email': email,
+                    'mobileNumber': phone_number,
+                }
+                return render(request, 'citizenAccount/citizenRegister.html', context)
 
         except OTPVerification.DoesNotExist:
-            messages.error(request, "Invalid OTP. Please try again.")
-            return redirect('citizenRegisterAccount')
 
-    return render(request, 'OTPScreen/OTPScreenRegistration.html', {'error': 'Invalid request method.'})
+            messages.error(request, "Invalid OTP. Please try again.")
+            context = {
+                'firstName': first_name,
+                'lastName': last_name,
+                'email': email,
+                'mobileNumber': phone_number,
+            }
+            return render(request, 'citizenAccount/citizenRegister.html', context)
+    
+    return render(request, 'OTPScreenOTPScreenRegistration.html', {'error': 'Invalid request method.'})
